Initial Website Structure and Recipe Links
In this commit, I've established the initial structure of the website. I've introduced a new directory called "recipes" to organise the recipes that will be featured on the website. Inside this directory, I've added the three recipes that are part of the website's content.

Furthermore, I've created links to these recipes on the index.html page to ensure easy navigation.