Added the Beef Stroganoff recipe to beef-stroganoff.html and modified the index.html file
In this commit I've added the complete Beef Stroganoff recipe, including the ingredients, instructions, and a high resolution image of the final dish.
I've also added an image to the index.html file and arranged the recipe links in an unordered list to create a better structure for the website.